Create README.md
Online Exam portal is just like quiz app it is designed by using the basic concepts of java such as classes objects abstraction interfaces the functionalities of this online exam portal project are we have login, signup, My profile, Exam, About and exit. In exam we are having 3 subjects science, English, maths in each subject we have 10questions for each question 2marks after the completion of exam we get score and the summary of our exam.